fix_partial_CDS function impremented
fix pseudo-partial CDS predicted at the end of the contig, but is likely to be intact.
The one that meet the following conditions will be retained
    - Starts at the coordinate 1 (+ strand) or at the end of contig (- strand)
    - 5'-end is missing and 3'-end is intact (+ strand) or the opposited case (- strand)
    - length is multiple of 3 and > min_length
    - the CDS has significant protein hit.

partial_flag 10, strand +, len>=500
partial_flag 01, strand -, len>=500

if the first 3 nuleotides are in [ATG, GTG, TTG], the CDS is fixed as intact. Otherwise, it is annotated as misc_feature